KAFKA-18862: Consolidate shared heartbeat request manager tests
